Rename SearchPagination into Pagination and move it under common
