+++ src/common/_common.scss
@@ -1,0 +1,1 @@
+@import './pagination/pagination';
