+++ src/domain/search/_search.scss
@@ -2,5 +2,4 @@
 @import './filters/searchFilters';
 @import './results/searchListResults';
 @import './results/searchMapResults';
-@import './pagination/searchPagination';
 @import './mapToggle/searchMapToggle';
